moved Admin to User folder (nested inside as a type of User).
