+++ rss-client/src/app/Earnpoints/components/earnpoints-overview-page/earnpoints-overview-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { BreadcrumbService } from 'src/app/services/breadcrumb.service';
+import { BreadcrumbService } from 'src/app/Common/services/breadcrumb.service';
 import { AppComponent } from 'src/app/app.component';
 
 @Component({
