+++ rss-client/src/app/User/Admin/admin-inventory/admin-inventory.component.html
