+++ rss-client/src/app/User/user.module.ts
@@ -5,6 +5,9 @@ import { LoginPageComponent } from './components/login-page/login-page.component
 import { AccountSettingsPageComponent } from './components/account-settings-page/account-settings-page.component';
 import { AccountPageComponent } from './components/account-page/account-page.component';
 import { AccountService } from './services/account.service';
+import { AdminPageComponent } from './Admin/admin-page/admin-page.component';
+import { AdminQuizComponent } from './Admin/admin-quiz/admin-quiz.component';
+import { AdminInventoryComponent } from './Admin/admin-inventory/admin-inventory.component';
 import { SESSION_STORAGE } from 'ngx-webstorage-service';
 import { USER_SERVICE_STORAGE, UserService } from './services/user.service';
 
@@ -14,7 +17,10 @@ import { USER_SERVICE_STORAGE, UserService } from './services/user.service';
   declarations: [
     LoginPageComponent,
     AccountSettingsPageComponent,
-    AccountPageComponent
+    AccountPageComponent,
+    AdminPageComponent,
+    AdminQuizComponent,
+    AdminInventoryComponent
   ],
   imports: [
     CommonModule,
