+++ rss-client/src/app/app-inventory/app-inventory.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { AddItemComponent } from '../app-inventory/components/add-item/add-item.component';
 import { InventoryListComponent } from '../app-inventory/components/inventory-list/inventory-list.component';
-import { AdminInventoryComponent } from '../Admin/admin-inventory/admin-inventory.component';
+import { AdminInventoryComponent } from 'src/app/User/Admin/admin-inventory/admin-inventory.component';
 
 export const inventoryRoutes: Routes = [
     {
