+++ rss-client/src/app/app-inventory/models/product.service.spec.ts
