+++ rss-client/src/app/app.component.ts
@@ -1,9 +1,9 @@
-import { QuizService } from './services/quiz.service';
+import { QuizService } from 'src/app/Quiz/service/quiz.service';
 import { Component, OnInit, Inject } from '@angular/core';
 import { User } from './User/models/user';
 import { Subscription } from 'rxjs';
 import { UserService } from './User/services/user.service';
-import { CheaterService } from './services/cheater.service';
+import { CheaterService } from 'src/app/Quiz/service/cheater.service';
 
 @Component({
   selector: 'app-root',
