--- rss-client/src/app/interfaces/cart-item.model.spec.ts
@@ -1,7 +1,0 @@
-import { CartItem } from './cart-item.model';
-
-describe('CartItem', () => {
-  it('should create an instance', () => {
-    expect(new CartItem()).toBeTruthy();
-  });
-});
