--- rss-client/src/app/interfaces/cart.model.spec.ts
@@ -1,7 +1,0 @@
-import { Cart } from './cart.model';
-
-describe('Cart', () => {
-  it('should create an instance', () => {
-    expect(new Cart()).toBeTruthy();
-  });
-});
