--- rss-client/src/app/interfaces/product.model.spec.ts
@@ -1,7 +1,0 @@
-import { Product } from './product.model';
-
-describe('Product', () => {
-  it('should create an instance', () => {
-    expect(new Product()).toBeTruthy();
-  });
-});
