--- rss-client/src/app/interfaces/product.model.ts
@@ -1,12 +1,0 @@
-export class Product {
-    id: number;
-    name: string;
-    description: string;
-    brand: string;
-    model: string;
-    category: string;
-    image: string;
-    quantity: number;
-    unitPrice: number;
-    color: string;
-}
