--- rss-client/src/app/interfaces/quiz.ts
@@ -1,13 +1,0 @@
-export interface Quiz {
-  quizId: number;
-  quizTopic: string;
-  quizDescription: string;
-  //Group 2 change
-  quizTotalPoints: number;
-  creatorEmail: string;
-  subjectId: number;
-  subject: {
-    subjectId: number;
-    subjectName: string;
-  };
-}
