--- rss-client/src/app/interfaces/quizSubmit.ts
@@ -1,18 +1,0 @@
-export interface QuizSubmit {
-  userScoreId: number;
-  userEmail: string;
-  userId: number;
-  userScore: 0;
-  submitDate: string;
-  quizId: number;
-  quiz: {
-    quizId: number;
-    quizTopic: string;
-    quizDescription: string;
-    creatorEmail: string;
-    subject: {
-      subjectId: number;
-      subjectName: string;
-    };
-  };
-}
