--- rss-client/src/app/interfaces/subject.ts
@@ -1,4 +1,0 @@
-export interface Subject {
-  subjectId: number;
-  subjectName: string;
-}
