--- rss-client/src/app/services/cart.service.spec.ts
@@ -1,20 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { CartService } from './cart.service';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-
-describe('CartService', () => {
-  let service: CartService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [],
-      imports: [HttpClientTestingModule],
-      providers: [CartService],
-    });
-    service = TestBed.inject(CartService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
