--- rss-client/src/app/services/image.service.spec.ts
@@ -1,9 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-describe('InventoryService', () => {
-  beforeEach(() => {});
-
-  // it('should be created', () => {
-  //   expect(service).toBeTruthy();
-  // });
-});
